+++ Configuration/Extbase/Persistence/Classes.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+return [
+    \Slavlee\PopupPower\Domain\Model\PopupContent::class => [
+        'tableName' => 'pages',
+        'recordType' => \Slavlee\PopupPower\Registry\PopupRegistry::PAGETYPES_POPUP_CONTENT
+    ],
+];