+++ Configuration/Icons.php
@@ -3,11 +3,12 @@
 declare(strict_types=1);
 
 return [
-    // Icon identifier
     'web-popuppower-dashboard' => [
-        // Icon provider class
         'provider' => \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
-        // The source SVG for the SvgIconProvider
+        'source' => 'EXT:popup_power/Resources/Public/Icons/Extension.svg',
+    ],
+    'pagetype-popup-content' => [
+        'provider' => \TYPO3\CMS\Core\Imaging\IconProvider\SvgIconProvider::class,
         'source' => 'EXT:popup_power/Resources/Public/Icons/Extension.svg',
     ],
 ];