+++ ext_tables.php
@@ -1,0 +1,8 @@
+<?php
+
+defined('TYPO3') || die();
+
+(static function () {
+    $obj = new \Slavlee\PopupPower\Bootstrap\ExtTables();
+    $obj->invoke();
+})();